Reformat VERBOSE to allow passing any level to sd-launch-pbs
VERBOSE in the PBS file was simply set to "true". Because of this we could only
pass args.verbose=0 or args.verbose=2 to sd-launch-pbs. VERBOSE is now set to
args.verbose and the command-line option {"", -v, -vv} is rebuilt in the PBS
script based on the value of args.parse.